Extract sockets to own file
